split the code #36
