+++ client/src/Components/Layouts/Navbar-v2/NavMenu.js
@@ -1,92 +1,8 @@
 import React from "react"
-import { Link } from "react-router-dom"
-import styled from "styled-components"
-import { sizes } from "../../Generic-helpers/variables"
+import { MenuLink, MenuSlider, MenuItem, MenuList } from "./StyledElm"
 import { ReactComponent as Close } from "../../../assets//images/baseline-close-24px.svg"
+import DropDownList from "./DropDownList"
 
-const MenuLink = styled(Link)`
-  width: 100%;
-  color: var(--black);
-  background-color: var(--whiet);
-  padding: 5px;
-  font-size: ${sizes.xl};
-  text-decoration: none;
-  &:active {
-    color: var(--primary);
-  }
-  &.active {
-    color: var(--primary);
-  }
-`
-const MenuItem = styled.li`
-  width: 90%;
-  border-top: 2px solid #c4c4c4;
-  list-style: none;
-  padding: 10px 0;
-  &.sub-menu-dropdown {
-    border-top: none !important;
-  }
-`
-const MenuList = styled.ul`
-  width: 100%;
-`
-const MenuSlider = styled.div`
-  display: inline-block;
-  &.show {
-    transform: translate3d(0vw, 0, 0);
-    overflow: hidden;
-  }
-  &.hide {
-    transform: translate3d(-100vw, 0, 0);
-  }
-  transition: transform 0.5s ease;
-  width: 100vw;
-  height: 100vh;
-  background-color: hsla(0, 0%, 100%, 1);
-  position: fixed;
-  top: 50px;
-  left: 0;
-  z-index: 1000;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  .nav-index-offset {
-    position: absolute;
-    right: 40px;
-    top: 0;
-    width: ${sizes.xxxxl};
-    height: ${sizes.xxxxl};
-    fill: var(--primary);
-    cursor: pointer;
-  }
-`
-const DropDownList = props => {
-  return (
-    <MenuList className="sub-menu">
-      <MenuItem className="sub-menu-dropdown">
-        <MenuLink to="/dashboard" onClick={props.handleMouseDown}>
-          My dashboard
-        </MenuLink>
-      </MenuItem>
-      <MenuItem className="sub-menu-dropdown">
-        <MenuLink to={`/profile/${props.profileId}`} onClick={props.handleMouseDown}>
-          View Profile{" "}
-        </MenuLink>
-      </MenuItem>
-      <MenuItem className="sub-menu-dropdown">
-        <MenuLink to="/edit-profile" onClick={props.handleMouseDown}>
-          Edit Profile
-        </MenuLink>
-      </MenuItem>
-      <MenuItem className="sub-menu-dropdown">
-        <MenuLink to="/postarticles" onClick={props.handleMouseDown}>
-          Create Article
-        </MenuLink>
-      </MenuItem>
-    </MenuList>
-  )
-}
 const NavMenu = props => {
   const { toggle, userName } = props
   const { user, isAuthenticated } = props.userAuthData
